haiku-generator/requirements.txt
accelerate==1.4.0
coqui-tts[server,ja]==0.25.3
cutlet==0.5.0
diffusers==0.32.2
fastapi==0.115.11
langchain-community==0.3.19
langchain-ollama==0.2.3
langchain==0.3.20
langgraph==0.3.5
minio==7.2.15
pydantic==2.10.6
pyfakefs==5.7.4
python-dotenv==1.0.1
sentencepiece==0.2.0
torch==2.6.0
uvicorn==0.34.0